requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dateutils==0.6.12
Django==4.0.4
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-nine==0.2.5
django-opensearch-dsl==0.2.0
djangorestframework==3.13.1
elasticsearch==7.17.3
elasticsearch-dsl==7.4.0
idna==3.3
opensearch-dsl==1.0.0
opensearch-py==1.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
